pyproject.toml
[tool.poetry]
name = "thermondo"
version = "0.1.0"
description = ""
authors = ["Ilhan Cetin <iilhancetin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
pytest = "^7.1.3"
pre-commit = "^2.20.0"
mypy = "^0.982"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
